refactor: move traits to App\Traits namespace
