+++ app/Models/Order.php
@@ -2,7 +2,7 @@
 
 namespace App\Models;
 
-use App\HasOrderStatus;
+use App\Traits\HasOrderStatus;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Support\Facades\DB;
