+++ app/Traits/HasSlug.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Traits;
 
 use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\Model;
